requirements_luiz.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
dataclasses-json==0.6.7
distro==1.9.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.19
langchain-community==0.0.38
langchain-core==0.1.53
langchain-experimental==0.0.58
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langsmith==0.1.147
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
openai==1.102.0
openai-whisper==20231117
orjson==3.11.3
packaging==23.2
pandas==2.2.2
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pynput==1.7.7
pyobjc-core==11.1
pyobjc-framework-ApplicationServices==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreText==11.1
pyobjc-framework-Quartz==11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
SQLAlchemy==2.0.43
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.11.0
torch==2.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
